Check directories exist before deleting in MCore post refinement cleanup (#297)

+++ MCore/MCore.cs
@@ -574,7 +574,8 @@ namespace MCore
                     }
 
                     foreach (var folder in AllProcessingFolders)
-                        Directory.Delete(folder, true);
+                        if (Directory.Exists(folder))
+                            Directory.Delete(folder, true);
 
                     Console.WriteLine("");
                 }
